Update 0.21 Gráfica implementada ( Apenas JOptionPane ) OP 1 corrigida ( Programa Final )

+++ src/main/NumericAlgorithm.kt
@@ -9,7 +9,7 @@ class NumericAlgorithm(description: String, title: String, author: String, versi
 
     fun shortenerEncode(inputNumber: Int): String {
 
-        var inputNumber: Int = inputNumber.toInt()
+        var inputNumber: Int = inputNumber
         var newBase = StringBuilder()
